requirements.txt
gunicorn
psycopg2
django-crispy-forms

django-heroku